Minor fixes
Divided socket handlers based on their usage. Added some env variable for testing.
