+++ client/src/pages/Game.js
@@ -47,7 +47,7 @@ const getCurrentSpymaster = ({ gameState: { turn, redSpy, blueSpy } }) => {
 const GamePage = ({ classes, state, setState, socket }) => {
   const { gameID } = useParams();
   const { gameState, player } = state;
-  const [timer, setTimer] = useState(20);
+  const [timer, setTimer] = useState(gameState !== undefined ? gameState.timer : undefined);
 
   const [tick] = useSound(tickSfx);
   const [correctGuess] = useSound(correctGuessSfx);
@@ -98,7 +98,7 @@ const GamePage = ({ classes, state, setState, socket }) => {
     const timerHandler = ({gameState, timer}) => {
       if (timer < 10)
         sounds.tick();
-        
+
       if (gameState)
         setState({ player: player, gameState: gameState });
