+++ server/.env.example
@@ -2,3 +2,5 @@ TEAM_MEMBERS="Aecio, Bonnie, Shums Kassam"
 CORS_ORIGIN=http://localhost:3000
 PORT=3001
 MONGODB_URL=mongodb+srv://<user>:<password>@<cluster>.mernf.mongodb.net/<dbname>?retryWrites=true&w=majority
+INIT_GAME_TIMER=30
+INIT_GUESS_CHANCE=2
