+++ server/socket/gameHandlers.js
@@ -1,38 +1,4 @@
-const gameController = require('../controllers/game')
-
-const disconnecting = (io, socket) => () => {
-  const gameList = Array.from(socket.rooms).filter(item => item!=socket.id)
-  console.log(`${socket.id} disconnecting`)
-  // remove the player from room
-  for (gameID of gameList) {
-    if (gameController.globalState[gameID] === undefined) {
-      console.log(`game (id: ${gameID}) does not exist.`);
-      continue;
-    }
-
-    const activePlayers = gameController.globalState[gameID].activePlayers
-    // remove the player from the active player list
-    delete activePlayers[socket.id]
-    if (Object.keys(activePlayers).length === 0)
-      delete gameController.globalState[gameID]
-  }
-};
-
-const disconnect = (io, socket) => () => {
-  console.log(`${socket.id} disconnected`);
-};
-
-const leave = (io, socket) => gameID => {
-  const room = gameController.globalState[gameID];
-  if (room) {
-    const activePlayers = room.activePlayers;
-    delete activePlayers[socket.id];
-    // remove the game if there are no active players
-    if (Object.keys(activePlayers).length === 0)
-      delete gameController.globalState[gameID];
-  }
-  io.emit('publicGames', gameController.getPublicGames());
-};
+const gameController = require('../controllers/game');
 
 const execute = (io, room, {method, params, word, callback}) => {
   const {id, gameEngine} = room;
@@ -112,14 +78,6 @@ const start = io => gameID => {
   io.emit('publicGames', gameController.getPublicGames());
 }
 
-// Propagate players chats & chat notifications to all in gameID
-const chat = io => (gameID, type, text, author) => {
-  if (process.env.NODE_ENV !== "production") {
-    console.log(`Received ${type} in game ${gameID} from ${author}: ${text}`);
-  }
-  io.to(gameID).emit("chat", type, text, author);
-};
-
 // Guesser's turn: Submit selected word
 const guesserNextMove = io => (gameID, player, word) => {
   const room = gameController.globalState[gameID];
@@ -171,21 +129,12 @@ const restartGame = io => gameID => {
   }
 }
 
-const refresh = io => () => {
-  io.emit('publicGames', gameController.getPublicGames());
-}
-
 module.exports = {
-  disconnecting,
-  disconnect,
-  leave,
   assign,
   unassign,
   start,
-  chat,
   guesserNextMove,
   spyNextMove,
   endTurn,
   restartGame,
-  refresh
 }
